3233806656983616238
ManifestID: 3233806656983616238
Date: 2 June 2022 – 16:52:29 UTC
Version: U42-509629

+++ Assembly-CSharp/MinionGroupProber.cs
@@ -19,7 +19,7 @@ public class MinionGroupProber : KMonoBehaviour, IGroupProber
 	{
 		base.OnPrefabInit();
 		MinionGroupProber.Instance = this;
-		this.cells = new Dictionary<object, int>[Grid.CellCount];
+		this.cells = new Dictionary<object, short>[Grid.CellCount];
 	}
 
 	private bool IsReachable_AssumeLock(int cell)
@@ -28,17 +28,17 @@ public class MinionGroupProber : KMonoBehaviour, IGroupProber
 		{
 			return false;
 		}
-		Dictionary<object, int> dictionary = this.cells[cell];
+		Dictionary<object, short> dictionary = this.cells[cell];
 		if (dictionary == null)
 		{
 			return false;
 		}
 		bool flag = false;
-		foreach (KeyValuePair<object, int> keyValuePair in dictionary)
+		foreach (KeyValuePair<object, short> keyValuePair in dictionary)
 		{
 			object key = keyValuePair.Key;
-			int value = keyValuePair.Value;
-			KeyValuePair<int, int> keyValuePair2;
+			short value = keyValuePair.Value;
+			KeyValuePair<short, short> keyValuePair2;
 			if (this.valid_serial_nos.TryGetValue(key, out keyValuePair2) && (value == keyValuePair2.Key || value == keyValuePair2.Value))
 			{
 				flag = true;
@@ -129,7 +129,7 @@ public class MinionGroupProber : KMonoBehaviour, IGroupProber
 		return this.IsReachable(Grid.PosToCell(workable), workable.GetOffsets());
 	}
 
-	public void Occupy(object prober, int serial_no, IEnumerable<int> cells)
+	public void Occupy(object prober, short serial_no, IEnumerable<int> cells)
 	{
 		object obj = this.access;
 		lock (obj)
@@ -138,19 +138,19 @@ public class MinionGroupProber : KMonoBehaviour, IGroupProber
 			{
 				if (this.cells[num] == null)
 				{
-					this.cells[num] = new Dictionary<object, int>();
+					this.cells[num] = new Dictionary<object, short>();
 				}
 				this.cells[num][prober] = serial_no;
 			}
 		}
 	}
 
-	public void SetValidSerialNos(object prober, int previous_serial_no, int serial_no)
+	public void SetValidSerialNos(object prober, short previous_serial_no, short serial_no)
 	{
 		object obj = this.access;
 		lock (obj)
 		{
-			this.valid_serial_nos[prober] = new KeyValuePair<int, int>(previous_serial_no, serial_no);
+			this.valid_serial_nos[prober] = new KeyValuePair<short, short>(previous_serial_no, serial_no);
 		}
 	}
 
@@ -167,9 +167,9 @@ public class MinionGroupProber : KMonoBehaviour, IGroupProber
 
 	private static MinionGroupProber Instance;
 
-	private Dictionary<object, int>[] cells;
+	private Dictionary<object, short>[] cells;
 
-	private Dictionary<object, KeyValuePair<int, int>> valid_serial_nos = new Dictionary<object, KeyValuePair<int, int>>();
+	private Dictionary<object, KeyValuePair<short, short>> valid_serial_nos = new Dictionary<object, KeyValuePair<short, short>>();
 
 	private List<object> pending_removals = new List<object>();
